Use rsync for vagrant-libvirt sync_folder:
The NFS type for synced_folder requires nfsd to be
installed and running on a machine for vagrant up to
be successful. nfs was used so that folders were r/w
bidirectionally. I don't really think there is a need for
this bidirectional functionality. Moving to rsync should
make this work more universally.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>