compose: Get rid of REPO_TOP_LEVEL
Its completely unnecessary. Docker-compose handles relative files relative
to the project folder, which by default is the folder that contains the
docker-compose.yml file. This makes the docker-compose.yml easier to parse
for humans.

Verified with `docker-compose -f $file config`, in and out of vagrant, in and
out of the compose dir, and by running a vagrant run.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>