tf/setup: Use apt-get helper function
This helper function lets us avoid the need to specify the env and extra args
every single time we want to interact with packages. This uses apt-get and not apt
because apt prints the following warning message:

> # WARNING: apt does not have a stable CLI interface. Use with caution in scripts.

and there's no overwhelming reason to favor apt > apt-get, so lets not risk it.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>