tf/setup: Improve correctness of get_second_interface_from_bond0
I noticed that the slaves names in the bonding/slaves file was not guaranteed
to be in any specific order so added a sort call. I also wanted to future proof
the function in case there's ever a type that has more than 2 nics in the bond
by default by ensuring we use the second nic. Finally, I made the function
behave correctly if the script is being re-run and the bond has already been
broken up, this was useful while debugging breakages and fixing this script.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>