tf: Add local variable for worker_macs
This variable doesn't assume that port[1] == "eth0", instead it grabs the mac
for the port whos  name is eth0. It also keeps track of the mac(s) in an array
because I was messing with multiple workers at one point and the logic to
figure this out was tricky and I don't want to lose it.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>