vagrant: Install docker and docker-compose via setup.sh
Without this a user needs to run vagrant more than once the first time they
try this out, once to fetch the vagrant-docker-compose plugin and then again
to actually run through. Why go through that when we already have code for
install docker and docker-compose that we have to maintain seprately anyway
and leads to a better first experience too?

Another minor benefit is that we increase the code sharing between terraform
and vagrant's setup.sh, which is nice.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>