vagrant: Allow overriding of network details from environment
This is nice to be able to have a hybrid setup with some things in vagrant
and actual hardware as long as its all bridged up nicely.  We can pretend
that machine1 is going to be used and thus go through all the tink setup
for it but then boot our real machine.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>