compose: Move configurable env vars to .env (also for DRYness)
Lets make .env the central/source-of-truth for env configuration. This
means we don't have to deal with empty vars and default values all over the
docker-compose.yml file. This does not cause any change in treatment of
environment variables. Actual environment variables still supersede those set
in .env. The values in .env are thus just the default if left unspecified,
exactly like ${VAR:-default} in the docker-compose.yml file.

I also got rid of container configs that used the `environment` config to
inject the env vars so that the `command` can use it. Having the values
in the environment would only really be useful if we expect to `exec`
into the container and run some commands. I haven't needed to do that yet
so would rather avoid repetition until absoulutely necessary. We can just
have docker-compose inject the values directly into the `command` line.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>