setup.sh: Add a message when setup.sh is all done
It is hard to tell when setup.sh is done in vagrant without having to
read/parse the output to figure out if we're at the last step. This change
makes it obvious. Its not so necessary for terraform but I'd like to keep
both scripts about as similar as possible.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>