Update in-repo documentation
This commit updates the in-repo docs to align it with the
tinkerbell.org docs.

I've also replaced the instructions for deploying Tinkerbell
in the Kubernetes quick start guide with a link to the
tinkerbell.org Install guide, so we don't have the same info
in too many places.

I've also removed links to the example manifests, as they
were not ported to the central repo. And again, best not
to have the same info in too many places.

Signed-off-by: Michael Meier <mmeier1986@gmail.com>